refactor signup container, components and routes
