+++ client/plugins/0.jquery/jquery.js
@@ -9064,7 +9064,7 @@ jQuery.fn.andSelf = jQuery.fn.addBack;
 // Register as a named AMD module, since jQuery can be concatenated with other
 // files that may use define, but not via a proper concatenation script that
 // understands anonymous AMD modules. A named AMD is safest and most robust
-// way to register. Lowercase 0.jquery is used because AMD module names are
+// way to signup. Lowercase 0.jquery is used because AMD module names are
 // derived from file names, and jQuery is normally delivered in a lowercase
 // file name. Do this after creating the global so that if an AMD module wants
 // to call noConflict to hide this version of jQuery, it will work.
