+++ client/stylesheets/globals/pages.import.less
@@ -61,7 +61,6 @@
   max-width: 400px;
   z-index: 100;
   margin: 0 auto;
-  padding-top: 40px;
 }
 
 .lockscreen.middle-box {
