+++ client/stylesheets/style.css
@@ -4173,7 +4173,6 @@ body.modal-open {
   max-width: 400px;
   z-index: 100;
   margin: 0 auto;
-  padding-top: 40px;
 }
 .lockscreen.middle-box {
   width: 200px;
