+++ imports/api/email/methods.js
@@ -37,6 +37,7 @@ export const send = new ValidatedMethod({
             subject: `theLeader.io`,
             html: html
           };
+          console.log(`sending welcome email`);
           Email.send(options);
         }
